backend/requirements.txt
flask
PyPDF2
Pillow
requests
boto3
anthropic
fitz
flask_cors
gunicorn